feat: [IOCOM-874] Bottom menu on SEND message details, with "Support" action item (#5644)
## Short description
This PR adds the bottom menu component (non-sticky) on the SEND message
details, with the working "Support" action item, for the new DS.
|Bottom menu|Support item|
|-|-|
|![Simulator Screenshot - iPhone 15 - 2024-03-28 at 10 30
25](https://github.com/pagopa/io-app/assets/5150343/5fd67ee8-dc52-4d2c-b915-c35760c780da)|![Simulator
Screenshot - iPhone 15 - 2024-03-28 at 10 30
29](https://github.com/pagopa/io-app/assets/5150343/0b589cf5-5ee6-4802-89a8-e824f463c278)|

## List of changes proposed in this pull request
- New `MessageFooter` component that references shared
`ContactsListItem` to handle the "Support" action item
- Old deprecated component renamed to `LegacyMessageFooter`

## How to test
Using the io-dev-api-server, generate some SEND messages and check that
the menu is displayed properly, with a working "Support" action.
