+++ ts/features/pn/components/__test__/MessageFooter.test.tsx
@@ -1,78 +1,33 @@
-import * as React from "react";
+import React from "react";
 import { createStore } from "redux";
 import { applicationChangeState } from "../../../../store/actions/application";
 import { preferencesDesignSystemSetEnabled } from "../../../../store/actions/persistedPreferences";
 import { appReducer } from "../../../../store/reducers";
 import { renderScreenWithNavigationStoreContext } from "../../../../utils/testWrapper";
-import { UIMessageId } from "../../../messages/types";
+import PN_ROUTES from "../../navigation/routes";
 import { MessageFooter } from "../MessageFooter";
-import * as standardPayments from "../../../messages/store/reducers/payments";
-import * as payments from "../../store/reducers/payments";
+import { ServiceId } from "../../../../../definitions/backend/ServiceId";
 
 describe("MessageFooter", () => {
-  beforeEach(() => {
-    jest.resetAllMocks();
-    jest.clearAllMocks();
-    jest
-      .spyOn(standardPayments, "canNavigateToPaymentFromMessageSelector")
-      .mockReturnValue(true);
-  });
-  it("should match snapshot for cancelled PN notification", () => {
-    jest
-      .spyOn(payments, "paymentsButtonStateSelector")
-      .mockReturnValue("visibleEnabled");
-    const component = renderScreen(true);
-    expect(component.toJSON()).toMatchSnapshot();
-  });
-  it("should match snapshot for hidden button", () => {
-    jest
-      .spyOn(payments, "paymentsButtonStateSelector")
-      .mockReturnValue("hidden");
-    const component = renderScreen();
-    expect(component.toJSON()).toMatchSnapshot();
-  });
-  it("should match snapshot for visibleLoading button", () => {
-    jest
-      .spyOn(payments, "paymentsButtonStateSelector")
-      .mockReturnValue("visibleLoading");
-    const component = renderScreen();
-    expect(component.toJSON()).toMatchSnapshot();
-  });
-  it("should match snapshot for visibleEnabled button", () => {
-    jest
-      .spyOn(payments, "paymentsButtonStateSelector")
-      .mockReturnValue("visibleEnabled");
-    const component = renderScreen();
+  it("should match snapshot", () => {
+    const component = renderComponent();
     expect(component.toJSON()).toMatchSnapshot();
   });
 });
 
-const renderScreen = (
-  isCancelled: boolean = false,
-  messageId: UIMessageId = "01HRAAFS3VJAAKWKV8NM8Z6CPQ" as UIMessageId,
-  maxVisiblePaymentCount: number = 5
-) => {
+const renderComponent = () => {
   const initialState = appReducer(undefined, applicationChangeState("active"));
   const designSystemState = appReducer(
     initialState,
     preferencesDesignSystemSetEnabled({ isDesignSystemEnabled: true })
   );
   const store = createStore(appReducer, designSystemState as any);
-  const mockRef = {
-    current: jest.fn()
-  };
 
   return renderScreenWithNavigationStoreContext(
     () => (
-      <MessageFooter
-        messageId={messageId}
-        maxVisiblePaymentCount={maxVisiblePaymentCount}
-        isCancelled={isCancelled}
-        payments={undefined}
-        presentPaymentsBottomSheetRef={mockRef}
-      />
+      <MessageFooter serviceId={"01HT25YR72A8N42AJ0TEKAB2V7" as ServiceId} />
     ),
-    "DUMMY",
+    PN_ROUTES.MESSAGE_DETAILS,
     {},
     store
   );
